# Parallel and optimized overlay builds