--- a/internal/pkg/config/util.go
+++ b/internal/pkg/config/util.go
@@ -1,0 +1,5 @@
+package config
+
+func BoolP(p *bool) bool {
+	return p != nil && *p
+}