@anderbubble 2024-12-21
Looks like something about BuildOverlay is not thread-safe. If I build a single node's overlays, it's fine.

```
$ sudo wwctl overlay build tn1
Building system overlays for tn1: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Created image for overlay tn1/[wwinit wwclient fstab hostname ssh.host_keys issue resolv udev.netname systemd.netname ifcfg NetworkManager debian.interfaces wicked ignition]: /var/lib/warewulf/provision/overlays/tn1/__SYSTEM__.img
Compressed image for overlay tn1/[wwinit wwclient fstab hostname ssh.host_keys issue resolv udev.netname systemd.netname ifcfg NetworkManager debian.interfaces wicked ignition]: /var/lib/warewulf/provision/overlays/tn1/__SYSTEM__.img.gz
Building runtime overlays for tn1: [hosts, ssh.authorized_keys, syncuser]
WARN   : Container is not defined for node: /etc/group
WARN   : Container is not defined for node: /etc/passwd
Created image for overlay tn1/[hosts ssh.authorized_keys syncuser]: /var/lib/warewulf/provision/overlays/tn1/__RUNTIME__.img
Compressed image for overlay tn1/[hosts ssh.authorized_keys syncuser]: /var/lib/warewulf/provision/overlays/tn1/__RUNTIME__.img.gz
```

But if I do a group I get an error.

```
$ sudo wwctl overlay build tn[1-10]
Building system overlays for tn1: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn10: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn2: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn3: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn4: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn5: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn6: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn7: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building runtime overlays for tn1: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn2: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn4: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn6: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn3: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn5: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn10: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn7: [hosts, ssh.authorized_keys, syncuser]
Building system overlays for tn8: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building system overlays for tn9: [wwinit, wwclient, fstab, hostname, ssh.host_keys, issue, resolv, udev.netname, systemd.netname, ifcfg, NetworkManager, debian.interfaces, wicked, ignition]
Building runtime overlays for tn8: [hosts, ssh.authorized_keys, syncuser]
Building runtime overlays for tn9: [hosts, ssh.authorized_keys, syncuser]
WARN   : Container is not defined for node: /etc/group
WARN   : Container is not defined for node: /etc/passwd
Created image for overlay tn9/[hosts ssh.authorized_keys syncuser]: /var/lib/warewulf/provision/overlays/tn9/__RUNTIME__.img
Compressed image for overlay tn9/[hosts ssh.authorized_keys syncuser]: /var/lib/warewulf/provision/overlays/tn9/__RUNTIME__.img.gz
ERROR: some overlays failed to be generated: could not build system overlays [wwinit wwclient fstab hostname ssh.host_keys issue resolv udev.netname systemd.netname ifcfg NetworkManager debian.interfaces wicked ignition] for node tn1: failed to generate files for overlay tn1/[wwinit wwclient fstab hostname ssh.host_keys issue resolv udev.netname systemd.netname ifcfg NetworkManager debian.interfaces wicked ignition]: failed to build overlay working directory: failed to render template warewulf/config.ww: could not parse template warewulf/config.ww: template: pattern matches no files: `warewulf/config.ww`
```