@anderbubble 2024-12-21
> something about BuildOverlay is not thread-safe

Probably this:

https://github.com/warewulf/warewulf/pull/1604/files#diff-be78947906f61c3025ba40b304a2f86ae28dad0ac9628fe8ecc9d03d679d2ec9L258