@anderbubble 2025-01-05
@griznog I would really appreciate your review on this one. I think it's all working as before; but with a new `--workers` option that defaults to the number of CPUs in the system; but testing it is making me think that overlay building is far too slow, so I want to profile it to see what's taking all the time.

That said, let me know if anything outright _breaks_ here, too.