@anderbubble 2025-01-05
I think I found the performance issue: each overlay build does a full `node.FindAllNodes`, because the overlay has access to all nodes. I think we could _drastically_ improve performance of overlay build by generating that list once and passing it in, rather than generating it for each overlay build. Along the way I'm writing a benchmark test so that this is easier to see in the future.