@anderbubble 2025-01-06
I added a benchmark test that builds overlays for 1000 nodes in a 1000-node cluster. Using this test, I located performance problems with overlay build (separate from serial vs parallel execution) and was able to take a ~13 minute overlay build down to ~13 seconds.

Before my performance fixes:

```
[janderson@rocky main]$ go test -timeout 30m github.com/warewulf/warewulf/internal/app/wwctl/overlay/build -bench Benchmark_Overlay_Build
goos: linux
goarch: arm64
pkg: github.com/warewulf/warewulf/internal/app/wwctl/overlay/build
Benchmark_Overlay_Build-8              1        773226980792 ns/op
PASS
ok      github.com/warewulf/warewulf/internal/app/wwctl/overlay/build   773.249s
```

After my performance fixes:

```
[janderson@rocky main]$ go test -timeout 30m github.com/warewulf/warewulf/internal/app/wwctl/overlay/build -bench Benchmark_Overlay_Build
goos: linux
goarch: arm64
pkg: github.com/warewulf/warewulf/internal/app/wwctl/overlay/build
Benchmark_Overlay_Build-8              1        13284022172 ns/op
PASS
ok      github.com/warewulf/warewulf/internal/app/wwctl/overlay/build   13.303s
```