@anderbubble 2025-01-06
It's not _quite_ as fast on the command-line, and I'm not sure why; but it's still way faster than it was.

```
# time wwctl overlay build tn[1-999] >/dev/null

real	1m1.311s
user	6m10.299s
sys	0m40.097s
```