@anderbubble 2025-01-06
To the reviewer:

This PR looks significantly bigger than it really is because I adjusted testenv to support "benchmarks" rather than just "tests". This incentivized a change to testenv's interface (an improvement regardless, imo), so I had to update the other test code to compensate.

Those changes should be confined to a single commit, so you can review them together for simplicity: https://github.com/warewulf/warewulf/pull/1604/commits/afda67c5ccf533705dcce25730298e8e11b09a70